refactor(hardware-correlator): drop the analysis scripts and the instrumented example
The offline analysis scripts, the pinned reproduction environment, the
vendored GNSSM2SDR snapshot and the software-control experiment were the
evidence trail of issue #107, not part of the receiver. The vendored
GNSSM2SDR change (a channel's confirmed arm sample) moves to GNSSM2SDR.jl.

The example is now a plain hardware-correlator run on the LiteX-M2SDR:
raw stream, device, link, receive, status lines, no warm-up and no
instrumentation, so it shows what the receiver does on its own.
