--- examples/analysis/repro_hwfix12/.gitignore
@@ -1,2 +1,0 @@
-/deps/
-/runs/
