--- examples/analysis/repro_hwfix12/Project.toml
@@ -1,17 +1,0 @@
-[deps]
-Acquisition = "d4bbf60b-102b-5ffb-8f97-a7ea5817e69f"
-FFTW = "7a1cc6ca-52ef-59f5-83cd-3a7055c09341"
-GNSSDecoder = "bf6b1ec8-12ad-4501-972c-3dec0d675c2b"
-GNSSM2SDR = "8a4c37e1-6d2b-4f9a-b3c5-1e7d90f24aa8"
-GNSSReceiver = "c84cffac-6943-4fe6-9449-f6d23817af8b"
-GNSSSignals = "52c80523-2a4e-5c38-8979-05588f836870"
-Geodesy = "0ef565a4-170c-5f04-8de2-149903a85f3d"
-JLD2 = "033835bb-8acc-5ee8-8aae-3f567f8a3819"
-PipeChannels = "316ea068-b8d4-4418-b34f-158552ec34f5"
-PositionVelocityTime = "52ec0b5e-ee36-11ea-211c-3d844fc5682d"
-Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
-SignalChannels = "59ec5e11-ed95-4167-ad38-35055440f478"
-StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
-Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
-Tracking = "10b2438b-ffd4-5096-aa58-44041d5c8f3b"
-Unitful = "1986cc42-f94f-5a68-af5c-568840ba703d"
