--- examples/analysis/repro_hwfix12/hardware.sha256
@@ -1,6 +1,0 @@
-3eb45748b9a3c255bfeb6a7581dc8c790ff305c615f4c2915ce89554d08d3690  csr.csv
-0f25c2ddec5e639e322934b6af8d9f286a8b0b43f0947a62872ce71e9033cdd7  gnss_m2sdr_m2_x1_ch20_ant1.bin
-ef6cdaa7ef0ee48e7b5048282048cdf9d1a5c982c7c031298c466615857d2352  host.txt
-8b047460829662f0b1a900b21c584dc43dd020252b3f392e80d876f323937e25  m2sdr.ko
-e5dbc41184319db861fc6d283a27627c71d4cc7abb3e934b22b9e8f2c6effd48  m2sdr_record
-6a12bf1fa948dc581067cf8006e9d98545f07b8a637e49d5827d0aa352f747a3  recorder-libraries.txt
