--- examples/analysis/repro_hwfix12/verification/first-arm/timing.json
@@ -1,12 +1,0 @@
-{
-  "timing_records": 50379,
-  "duration_s": 100.155525208,
-  "compilation_ms": 8.410704,
-  "max_chunk_gap_ms": 29.79738,
-  "compilation_events": [
-    [
-      100.063096165,
-      8.410704
-    ]
-  ]
-}
