--- examples/hardware_correlator_position_fix.jl
@@ -1,1639 +1,0 @@
-# Live GPS L1 C/A position fix through the LiteX-M2SDR hardware correlator.
-#
-# The full GNSSReceiver pipeline — acquisition, decoding, PVT — runs on the CPU
-# off the raw DMA0 stream; correlation and the tracking NCOs run on the FPGA,
-# linked through GNSSReceiver's hardware-correlator interface (issue #107).
-#
-# Structure:
-#   1. JIT warm-up of the receive() path on synthetic samples.
-#   2. Raw stream: m2sdr_record → libuv pipe → SignalChannel (antenna 0 of the 2R2T
-#      words), with a producer-side sample count and a tap ring for the opening
-#      scan.
-#   3. One 32-PRN scan to prioritize the receiver's search order. The host↔device sample
-#      axis needs no calibration: the bank's counter is latched immediately
-#      before streaming starts and only counts samples DMA0 accepts, so the
-#      latch is exact up to buffers dropped in between — measured at 0 or 1
-#      sample (≤0.25 chips) across 36 runs. `slip_watch` re-anchors it during
-#      the run, which is where the mapping really does drift.
-#   4. receive(sdr, …) with DMA-streamed dumps (dump_source = :dma,
-#      epoch_period = 50 → ~2.2 ms feedback transport) until the PVT solve
-#      returns a fix; it then rides RUN_AFTER_FIX seconds and stops.
-#
-# Usage:
-#   julia -t 6,2 --project=. hardware_correlator_position_fix.jl [MAX_SECONDS] [RUN_AFTER_FIX]
-# (defaults 600 and 30; HWFIX_ACQ_EVERY / HWFIX_ASYNC set the rescan cadence and
-#  whether the scan runs off the processing task. One default-pool thread is for
-#  the acquisition worker, so give it at least three.)
-#
-# Bit-stream instrument (issue #107). Set HWFIX_BITLOG=<name> to log, next to
-# this script, every soft bit and every 1 ms filtered prompt on the way into
-# `decode`, plus the E/P/L accumulators, both Dopplers and the per-channel
-# record-continuity counters:
-#
-#   HWFIX_BITLOG=bitstream.log julia -t 6,2 --project=. \
-#       hardware_correlator_position_fix.jl 220 20
-#   julia analysis/subframes_softbits.jl bitstream.log 7      # real subframes?
-#   julia analysis/word_errors.jl        bitstream.log 7 993  # where the errors are
-#   julia analysis/walkoff.jl            bitstream.log 10     # E/L imbalance over time
-#
-# Unset (the default) it costs nothing: the wrapper's per-chunk hook returns
-# immediately. Logging goes through the `correlator_source` seam rather than
-# `receive`'s payload channel because that channel emits once per
-# `pvt_update_interval` while the bit buffer is reset every chunk — sampling it
-# sees ~1/50 of the bits. `advance_tracking!` runs after the fold and before
-# `decode` consumes the buffer, so what it sees IS what the decoder gets.
-#
-# FPGA-vs-CPU correlator probe (issue #107). Set HWFIX_XCORR=<name> to correlate
-# the *same raw samples* on the CPU, with the *same NCO parameters*, once per
-# hardware record, and log both prompts side by side:
-#
-#   HWFIX_XCORR=xcorr.log HWFIX_XCORR_PRNS=7,20 julia -t 6,2 --project=. \
-#       hardware_correlator_position_fix.jl 120 10
-#   julia analysis/fpga_vs_cpu.jl xcorr.log
-#
-# This is the direct measurement of what the hardware correlator loses: the
-# 200 s bit-error study (issue #107) put a ~3 % BER on the FPGA bit stream while
-# the pure software receiver validated full ephemerides off the same DMA0 stream
-# in the same minute, which leaves only the correlator itself. Both streams see
-# one input, so every shared error source — antenna, RF, DMA0, acquisition —
-# cancels, and the residual is the device's.
-#
-# Companion branches (this exact stack produced the first live fix 2026-08-01,
-# lat 50.768612° lon 6.072698° alt 271.2 m):
-#   GNSSM2SDR.jl feat/code-phase-anchor — immediate-CSR NCO updates, record
-#     dedup, O_NONBLOCK DMA reader; Tracking.jl hwfix/hardware-correlator-run —
-#     1 ms integrations, degenerate-record skips.
-#
-# Board bring-up (Jetson + LiteX M2SDR; redo after EVERY reboot):
-#   sudo insmod ~/litex_m2sdr/litex_m2sdr/software/kernel/m2sdr.ko
-#     (that build has DMA_BUFFER_COUNT=256 — the gateware descriptor table is
-#      256 deep; larger rings wrap it into stale-buffer corruption. Never
-#      modprobe: the installed module is an older single-device build.)
-#   sudo sysctl -w fs.pipe-max-size=67108864   # not persistent
-#   ~/litex_m2sdr/litex_m2sdr/software/user/m2sdr_rf \
-#     --sample-rate 4000000 --bandwidth 2500000 --rx-freq 1575420000 \
-#     --rx-gain 60 --tx-att 89
-#     (long flags only; a bare m2sdr_rf resets to 30.72M/2.4GHz defaults)
-#   Verify: m2sdr_record streams ~32 MB/s with ~0 adjacent-duplicate samples.
-#   MAXN power mode + jetson_clocks: the scan itself now overlaps tracking
-#   (asynchronous acquisition), but it still competes for cores with the fold
-#   and the CSR/DMA reader, so the clocks matter.
-
-using Printf
-using Statistics: mean, median, std
-using StaticArrays: SVector
-using Unitful
-using Unitful: Hz, ms, ustrip
-using Geodesy: LLAfromECEF, wgs84, ECEF
-
-using Acquisition: acquire
-using FFTW
-using GNSSSignals: GPSL1CA, get_code, get_code_frequency
-import Tracking
-using Tracking: ConventionalAssistedPLLAndDLL
-using SignalChannels: SignalChannel
-using GNSSReceiver
-using GNSSM2SDR
-using GNSSM2SDR: schedule!, load_code!, spacing_word, carrier_word, code_word,
-    sample_count, apply_status, GPS_L1_HZ, GPS_CA_CHIP_RATE, CA_CODE_LENGTH
-
-const FS_HZ = 4e6
-const FS = 4e6Hz
-# 2 ms chunks: per-chunk pipeline bookkeeping (ReceiverState rebuild, extract,
-# decode consume, PVT) is a fixed cost per chunk; at 4000-sample chunks the
-# receiver runs ~0.55x real time with 5 sats on this box, and rt<1x means
-# either sample loss (small pipe, old runs) or unbounded lag (big pipe). The
-# fold grid stays at 1 ms epochs inside advance_tracking!; only feedback
-# batching coarsens (~+2 ms tau, fine at 8 Hz PLL).
-const CHUNK = 8000                       # 2 ms
-const CSR_CSV = expanduser("~/gnss-m2sdr/build/gnss_m2sdr_m2_x1_ch20_ant1/csr.csv")
-const FRAC = 24
-const ACQ_COVERAGE = 25_000.0Hz          # one-sided; LO offset ~14 kHz + Doppler
-# Diagnostic floor for the opening scan's visibility report. It does not limit
-# later acquisition: a satellite can become visible after startup, and every
-# periodic scan must remain able to discover it. CFAR in receive() decides
-# detection; the hardware link enforces channel capacity.
-const RX_FLOOR_DBHZ = 30.0
-const MAX_SECONDS = length(ARGS) >= 1 ? parse(Float64, ARGS[1]) : 600.0
-const RUN_AFTER_FIX = length(ARGS) >= 2 ? parse(Float64, ARGS[2]) : 30.0
-# Periodic rescan cadence. Acquisition now runs off the processing task
-# (`acquire_async`), so a scan no longer stalls the chunk pipeline and a real
-# rescan cadence is affordable — the `t=` vs `rt=` gap in the progress line is
-# the direct measurement of that. `HWFIX_ASYNC=0` puts the search back inline
-# for comparison, which is what forced the earlier runs to disable the rescan
-# altogether (`acquire_every = 300 s`).
-const ACQ_EVERY = parse(Float64, get(ENV, "HWFIX_ACQ_EVERY", "30")) * u"s"
-const ACQ_ASYNC = get(ENV, "HWFIX_ASYNC", "1") == "1"
-# Which thread pool the chunk-processing task runs on (`receive`'s
-# `processing_threadpool`). `:interactive` keeps the fold clear of the
-# acquisition workers on the default pool; give Julia the
-# threads for it (`-t 6,3`: raw reader, DMA reader + NCO writer, and the fold).
-const PROC_POOL = Symbol(get(ENV, "HWFIX_PROC_POOL", "default"))
-
-const gpsl1 = GPSL1CA()
-# Bit-stream instrument (issue #107): unset means no logging, no cost.
-const BIT_LOG_NAME = get(ENV, "HWFIX_BITLOG", "")
-# FPGA-vs-CPU correlator probe (issue #107): likewise off unless named.
-const XCORR_LOG_NAME = get(ENV, "HWFIX_XCORR", "")
-# Restrict the probe to these PRNs (empty ⇒ every assigned satellite). One CPU
-# correlation per hardware record costs ~50 µs, so five satellites at 1 kHz add
-# ~0.25 s of CPU per second of signal to the *processing task* — watch the
-# `t=` vs `rt=` gap and narrow this (or raise HWFIX_XCORR_EVERY) if it grows.
-const XCORR_PRNS =
-    Set(parse.(Int, filter(!isempty, split(get(ENV, "HWFIX_XCORR_PRNS", ""), ','))))
-const XCORR_EVERY = parse(Int, get(ENV, "HWFIX_XCORR_EVERY", "1"))
-const RAW_CHUNKS = Threads.Atomic{Int}(0)
-
-# Keep FFTW single-threaded. `Acquisition` already parallelises a scan across
-# PRNs on Julia threads, so FFTW threads inside each PRN's transforms only add
-# spinning barriers on an oversubscribed machine. Measured on the Orin with the
-# receiver's own plan (25 kHz coverage, 10 ms coherent, 5 rounds): 4 PRNs take
-# 0.48 s with one FFTW thread and 6.6 s with four; all 32 PRNs 1.7 s against
-# 38 s. The "55 s scans" of issue #107 were this setting, not the search.
-FFTW.set_num_threads(1)
-
-# ── 1. JIT warm-up ────────────────────────────────────────────────────────────
-# The warm-up must compile the SAME specializations the live call uses: the
-# HARDWARE advance_tracking!/fold path (via an AbstractHardwareCorrelatorSDR
-# source with real dumps), the ConventionalAssistedPLLAndDLL track state, the
-# PRN-restricted acquisition and the extract payload. A software-source warm-up
-# with different kwargs leaves 25-110 s of JIT to happen live, during which the
-# chunk backlog grows and every tracking loop is born open.
-using StaticArrays: SVector
-using PipeChannels: PipeChannel
-using Tracking: CorrelatorOutput, EarlyPromptLateCorrelator
-using GNSSReceiver: CorrelatorDump, NCOUpdate, AbstractHardwareCorrelatorSDR, epoch_strobe
-
-# ── FPGA ↔ CPU correlator probe ───────────────────────────────────────────────
-# Correlate the same raw samples the FPGA correlated, over the same span, with
-# the same NCO parameters, and log both prompts per record.
-#
-# The alignment is the whole trick. A hardware record covers device samples
-# `[sample_index - integrated_samples, sample_index)` and reports the code NCO's
-# own phase at its end; the raw stream is the same samples, offset by the
-# constant `sdr.device_origin` (the run's origin calibration is what pins it).
-# So the probe rings the raw chunks by device index and, for each record, slices
-# exactly that span back out. Nothing is resampled or re-timed: if the two
-# prompts disagree, the disagreement is in the correlator.
-#
-# Three things are deliberately NOT taken from the host's model of the device:
-#
-#   * The code phase comes off the wire (`CorrelatorDump.code_phase`), so the
-#     CPU replica sits where the device's replica actually sat, not where the
-#     host thinks it steered it.
-#   * A slow CPU-side DLL (`delta`, chips) rides on top of it. An exact device
-#     axis leaves it at zero; a residual is the host↔device sample-mapping error
-#     in chips, measured rather than assumed — and it keeps the CPU reference on
-#     the peak while that is measured, so the comparison stays fair.
-#   * The mixing sign is voted on at the first record rather than inherited from
-#     Tracking's convention: getting it wrong costs the CPU side all of its
-#     correlation gain, and nothing else in the output would say so.
-#
-# The carrier *frequency* is shared on purpose — both sides mix with the Doppler
-# the loop commanded. A carrier NCO that does not honour that word then shows up
-# offline as a phase ramp on the FPGA prompt against a stationary CPU one.
-#
-# The kernels live in `analysis/cpu_correlator.jl` so they can be pinned against
-# an analytically known injected signal without a board; see
-# `analysis/selftest_cpu_correlator.jl`.
-include(joinpath(@__DIR__, "analysis", "cpu_correlator.jl"))
-using .CpuCorrelator: SampleRing, Workspace, push_chunk!, covers, correlate_epl,
-    search_offset, accumulate_offsets!, peak_offset
-
-# The coarse search slides the code replica ±this many whole samples (±16 chips
-# at 4 MHz), which covers any plausible residual of the origin calibration.
-const XCORR_SEARCH_SAMPLES = 64
-
-# A record waiting for its samples, with everything the comparison needs frozen
-# at the moment it was drained.
-#
-# Records reach the host *ahead* of the raw samples they were computed from: a
-# dump crosses DMA1 in a couple of milliseconds while the same samples are still
-# queued in the recorder pipe and the receiver's own sample channel. Correlating
-# only what the ring already holds would therefore throw away nearly every
-# record — so a record is parked here and picked up once the raw stream has
-# caught up with it, typically one or two chunks later.
-#
-# Freezing the Dopplers at drain time (rather than reading them when the samples
-# arrive) also keeps the CPU replica on the loop state that was actually driving
-# the device around that record.
-struct PendingRecord
-    channel::Int32
-    prn::Int32
-    sample_index::Int64
-    n::Int32
-    code_phase::Float64
-    late::ComplexF64
-    prompt::ComplexF64
-    early::ComplexF64
-    carrier_hz::Float64
-    code_doppler_hz::Float64
-    code_freq_hz::Float64
-    shift::Int32
-end
-
-mutable struct CpuReference
-    const io::IO
-    # Device sample index of host raw-sample 0, i.e. `sdr.device_origin` at the
-    # moment `receive` took the stream over. Constant for the run.
-    const origin::Int64
-    const fs::Float64
-    const ring::SampleRing
-    const ws::Workspace
-    const codes::Dict{Int,Vector{Float32}}
-    const prns::Set{Int}
-    const every::Int
-    # Records drained but not yet reached by the raw stream.
-    const pending::Vector{PendingRecord}
-    const max_pending::Int
-    # Per hardware channel.
-    const last_index::Vector{Int64}      # newest record seen, for dedup
-    const delta::Vector{Float64}         # CPU code-phase offset, chips
-    const anchor::Vector{Float64}        # where the coarse search put `delta`
-    const occupant::Vector{Int32}        # PRN the probe last coarse-aligned to
-    # The CPU's own carrier NCO, mirroring the device's: phase in cycles, the
-    # device sample it refers to, and the frequency running since then.
-    const carrier_cycles::Vector{Float64}
-    const carrier_ref::Vector{Int64}
-    const last_carrier_hz::Vector{Float64}
-    const since_search::Vector{Int}      # records since this channel was aligned
-    # The accumulating alignment search, per channel: how many records are still
-    # to be summed, over what margin, whether this cycle began at a handover,
-    # and the power over offsets so far.
-    const search_left::Vector{Int}
-    const search_margin::Vector{Int}
-    const search_fresh::Vector{Bool}
-    const search_acc::Vector{Vector{Float64}}
-    const seen::Vector{Int}
-    sign::Int                            # carrier mixing sign, 0 until voted
-    records::Int
-    skipped_old::Int                     # the ring rolled over the samples first
-    dropped_pending::Int                 # the queue overflowed
-    unaligned::Int                       # coarse search found no peak
-    max_wait::Int64                      # deepest record→sample lag seen, samples
-    cpu_seconds::Float64
-end
-
-# Alignment searching. `XCORR_SEARCH_RECORDS` records are accumulated per search
-# — one record's argmax over 129 offsets is frequently noise, and a search built
-# on it walks the reference off the satellite while reporting a plausible ratio.
-# After a search succeeds the reference is close, so the next one only has to
-# watch ±`XCORR_TRACK_SAMPLES`; a failure widens it back to the full window.
-# Re-searching every `XCORR_RESEARCH_RECORDS` is what makes reference alignment
-# observable across the run instead of assumed from one measurement at handover.
-const XCORR_SEARCH_RECORDS = 20
-const XCORR_TRACK_SAMPLES = 8
-const XCORR_RESEARCH_RECORDS = 5000
-# Peak-to-floor a search must reach to be believed, and how much better than the
-# offset it is already using a move has to be.
-const XCORR_PEAK_RATIO = 5.0
-const XCORR_MOVE_RATIO = 1.5
-
-function CpuReference(io::IO, origin::Integer, fs::Real, n_channels::Integer;
-                      ring_seconds = 0.5, prns = Set{Int}(), every::Integer = 1,
-                      max_pending::Integer = 8192)
-    # One field per line, named. Thirty positional arguments is a shape that
-    # silently mis-assigns whenever a field is inserted, and it did.
-    p = CpuReference(
-        io,                                                   # io
-        Int64(origin),                                        # origin
-        Float64(fs),                                          # fs
-        SampleRing(round(Int, ring_seconds * fs)),            # ring
-        Workspace(2 * CHUNK, XCORR_SEARCH_SAMPLES),           # ws
-        Dict{Int,Vector{Float32}}(),                          # codes
-        Set{Int}(prns),                                       # prns
-        Int(every),                                           # every
-        PendingRecord[],                                      # pending
-        Int(max_pending),                                     # max_pending
-        fill(typemin(Int64), n_channels),                     # last_index
-        zeros(n_channels),                                    # delta
-        zeros(n_channels),                                    # anchor
-        zeros(Int32, n_channels),                             # occupant
-        zeros(n_channels),                                    # carrier_cycles
-        fill(typemin(Int64), n_channels),                     # carrier_ref
-        zeros(n_channels),                                    # last_carrier_hz
-        zeros(Int, n_channels),                               # since_search
-        zeros(Int, n_channels),                               # search_left
-        fill(XCORR_SEARCH_SAMPLES, n_channels),               # search_margin
-        fill(false, n_channels),                              # search_fresh
-        [zeros(2 * XCORR_SEARCH_SAMPLES + 1) for _ = 1:n_channels],  # search_acc
-        zeros(Int, n_channels),                               # seen
-        0,                                                    # sign
-        0,                                                    # records
-        0,                                                    # skipped_old
-        0,                                                    # dropped_pending
-        0,                                                    # unaligned
-        Int64(0),                                             # max_wait
-        0.0,                                                  # cpu_seconds
-    )
-    # Compile the kernels here rather than during the first live chunk: the
-    # probe runs inside the processing task, and a JIT pause there is a chunk
-    # backlog and a burst of stale NCO commits. They are separate methods, so
-    # they have to be *run*, not merely referenced — over the still-zeroed ring,
-    # which correlates to nothing and costs a millisecond. Nothing durable is
-    # touched: the sign vote is `search_offset`'s return value, not its state.
-    let n = 400, code = xcorr_code_table!(p, 1)
-        correlate_epl(p.ws, p.ring, Int64(0), n; code, cp_start = 0.0,
-                      cps = 0.2558, freq = 1000.0, fs = p.fs, sign = 1, shift = 1)
-        search_offset(p.ws, p.ring, Int64(0), n; code, cp_start = 0.0,
-                      cps = 0.2558, freq = 1000.0, fs = p.fs,
-                      margin = XCORR_SEARCH_SAMPLES)
-    end
-    println(io, "# X prn ch sample_index n code_phase delta_chips ",
-                "re_Pf im_Pf abs_Lf abs_Ef re_Pc im_Pc abs_Lc abs_Ec ",
-                "carrier_doppler code_doppler")
-    println(io, "# A/R prn ch sample_index offset_samples offset_chips ",
-                "peak_over_median peak_over_centre offsets_searched")
-    println(io, "# S mixing_sign peak_over_median")
-    p
-end
-
-xcorr_code_table!(p::CpuReference, prn::Integer) = get!(p.codes, Int(prn)) do
-    Float32[get_code(gpsl1, i, prn) for i = 0:(CA_CODE_LENGTH-1)]
-end
-
-# Until the first usable vote, mix as Tracking does; every `A` line carries the
-# vote's evidence, so a run that never voted says so.
-_xcorr_sign(p::CpuReference) = p.sign == 0 ? 1 : p.sign
-
-xcorr_append!(p::CpuReference, samples, host_start::Integer) =
-    push_chunk!(p.ring, samples, p.origin + Int64(host_start))
-
-# Capture one drained record: freeze what the comparison will need and park it
-# until the raw stream reaches it.
-function xcorr_capture!(p::CpuReference, assignments, track_state, dump)
-    ch = Int(dump.channel)
-    (1 <= ch <= length(p.last_index)) || return p          # epoch strobe
-    dump.output.sample_index <= p.last_index[ch] && return p
-    p.last_index[ch] = dump.output.sample_index
-    assignment = assignments[ch]
-    isnothing(assignment) && return p
-    assignment.prn == Int(dump.prn) || return p
-    (isempty(p.prns) || assignment.prn in p.prns) || return p
-    isnan(dump.code_phase) && return p
-    n = Int(dump.output.integrated_samples)
-    (n <= 0 || n > length(p.ws.mixed)) && return p
-    p.seen[ch] += 1
-    (p.every > 1 && p.seen[ch] % p.every != 0) && return p
-
-    sat_states = Tracking.get_sat_states(track_state, assignment.group_key)
-    haskey(sat_states, assignment.prn) || return p
-    sat = sat_states[assignment.prn]
-    tracked = Tracking.get_signals(sat)[assignment.signal_index]
-    signal = Tracking.get_signal(tracked)
-    code_doppler = ustrip(Hz, uconvert(Hz, Tracking.get_code_doppler(sat)))
-    code_freq = ustrip(Hz, get_code_frequency(signal)) + code_doppler
-    # Whole-sample Early↔Late spacing, quantised exactly the way `_assign!`
-    # quantised it for the gateware, so the CPU taps sit where the device's do.
-    shift = max(1, round(Int, Tracking.get_early_late_sample_spacing(
-        Tracking.get_correlator(tracked), p.fs * Hz, code_freq * Hz) / 2))
-    acc = Tracking.get_accumulators(dump.output.correlator)   # [late, prompt, early]
-
-    if length(p.pending) >= p.max_pending
-        # The raw stream is not catching up (or has stalled). Drop the oldest —
-        # a diagnostic must never become the reason the receiver falls behind.
-        popfirst!(p.pending)
-        p.dropped_pending += 1
-    end
-    push!(p.pending, PendingRecord(
-        Int32(ch), Int32(assignment.prn), Int64(dump.output.sample_index), Int32(n),
-        dump.code_phase,
-        _xcorr_scalar(acc[1]), _xcorr_scalar(acc[2]), _xcorr_scalar(acc[3]),
-        ustrip(Hz, uconvert(Hz, Tracking.get_carrier_doppler(sat))),
-        code_doppler, code_freq, Int32(shift)))
-    p
-end
-
-# Correlate every parked record the raw stream has now reached, in order, and
-# log both sides. Records still ahead of the stream stay parked.
-function xcorr_drain_pending!(p::CpuReference)
-    keep = 0
-    for rec in p.pending
-        s0 = rec.sample_index - rec.n
-        status = covers(p.ring, s0, Int(rec.n))
-        if status === :future
-            keep += 1
-            p.pending[keep] = rec
-            p.max_wait = max(p.max_wait, rec.sample_index - p.ring.ending)
-            continue
-        elseif status === :old
-            p.skipped_old += 1
-            continue
-        end
-        xcorr_compare!(p, rec, s0)
-    end
-    resize!(p.pending, keep)
-    p
-end
-
-# One record: correlate the raw span the FPGA integrated over, at the code phase
-# the FPGA reported, with the Doppler the loop commanded — then log both sides.
-function xcorr_compare!(p::CpuReference, rec::PendingRecord, s0::Int64)
-    ch = Int(rec.channel)
-    n = Int(rec.n)
-    shift = Int(rec.shift)
-    cps = rec.code_freq_hz / p.fs
-    code = xcorr_code_table!(p, rec.prn)
-    # The dump's phase is the replica's at the end of the integration.
-    cp_start = rec.code_phase - n * cps + p.delta[ch]
-
-    # Mirror the device's carrier NCO instead of restarting at zero. Its phase
-    # runs continuously across records, so a reference that restarts each time
-    # differs from it by `carrier_doppler x record length` — tens of cycles per
-    # record at GPS Dopplers, aliasing to an effectively random offset. That
-    # leaves every magnitude intact and destroys every phase comparison, which
-    # is exactly the comparison this probe exists to make.
-    if p.carrier_ref[ch] == typemin(Int64) || p.occupant[ch] != rec.prn
-        p.carrier_cycles[ch] = 0.0
-    else
-        p.carrier_cycles[ch] = mod(
-            p.carrier_cycles[ch] +
-            p.last_carrier_hz[ch] * (s0 - p.carrier_ref[ch]) / p.fs, 1.0)
-    end
-    p.carrier_ref[ch] = s0
-    p.last_carrier_hz[ch] = rec.carrier_hz
-    phase0 = p.carrier_cycles[ch]
-
-    # A fresh occupant is coarse-aligned again. `delta` is not reset with it:
-    # most of it is the host↔device axis error, which every channel shares, so
-    # the previous occupant's value is the best starting point.
-    # The mixing sign is voted on once per run, on the first record compared:
-    # inheriting Tracking's convention and getting it wrong costs the CPU side
-    # all of its correlation gain, and nothing else in the output would say so.
-    # One record settles it — the wrong sign leaves 5 kHz of residual carrier
-    # inside a 1 ms integration, which is not a near thing.
-    if p.sign == 0
-        _, ratio, sign = search_offset(
-            p.ws, p.ring, s0, n; code, cp_start, cps, freq = rec.carrier_hz,
-            fs = p.fs, signs = (1, -1), margin = XCORR_SEARCH_SAMPLES, phase0)
-        p.sign = sign
-        @printf(p.io, "S %+d %.2f\n", sign, ratio)
-    end
-
-    # Start an alignment search: at every handover, and periodically after that.
-    fresh = p.occupant[ch] != rec.prn
-    if fresh
-        p.occupant[ch] = rec.prn
-        p.search_margin[ch] = XCORR_SEARCH_SAMPLES
-    end
-    p.since_search[ch] += 1
-    if p.search_left[ch] == 0 &&
-       (fresh || p.since_search[ch] >= XCORR_RESEARCH_RECORDS)
-        p.search_left[ch] = XCORR_SEARCH_RECORDS
-        p.search_fresh[ch] = fresh
-        p.since_search[ch] = 0
-        fill!(p.search_acc[ch], 0.0)
-    end
-
-    if p.search_left[ch] > 0
-        m = p.search_margin[ch]
-        acc = view(p.search_acc[ch], 1:(2m+1))
-        accumulate_offsets!(acc, p.ws, p.ring, s0, n; code, cp_start, cps,
-                            freq = rec.carrier_hz, fs = p.fs,
-                            sign = _xcorr_sign(p), margin = m, phase0)
-        p.search_left[ch] -= 1
-        if p.search_left[ch] == 0
-            offset, ratio, over_centre = peak_offset(acc, m)
-            # Move only on a peak that is both real and better than where the
-            # reference already sits. Anything else and a noisy scan would drag
-            # a perfectly aligned reference off the satellite — which is exactly
-            # what a single-record search did before it accumulated.
-            if ratio >= XCORR_PEAK_RATIO &&
-               (offset == 0 || over_centre >= XCORR_MOVE_RATIO)
-                p.delta[ch] += offset * cps
-                p.anchor[ch] = p.delta[ch]
-                cp_start += offset * cps
-                p.search_margin[ch] = XCORR_TRACK_SAMPLES
-            elseif ratio < XCORR_PEAK_RATIO
-                # Nothing found where we were looking: look everywhere again.
-                p.search_margin[ch] = XCORR_SEARCH_SAMPLES
-                p.search_fresh[ch] && (p.unaligned += 1)
-            end
-            @printf(p.io, "%s %d %d %d %d %.4f %.2f %.2f %d\n",
-                    p.search_fresh[ch] ? "A" : "R", rec.prn, ch,
-                    rec.sample_index, offset, offset * cps, ratio, over_centre,
-                    2m + 1)
-        end
-    end
-
-    late, prompt, early = correlate_epl(
-        p.ws, p.ring, s0, n; code, cp_start, cps, freq = rec.carrier_hz,
-        fs = p.fs, sign = _xcorr_sign(p), shift, phase0)
-
-    # Slow normalised early-late envelope DLL on the CPU side. Deliberately
-    # sluggish, and bounded to the coarse search's anchor: it exists to hold the
-    # reference on the peak and to *measure* the residual axis error, not to
-    # track the satellite — the FPGA is doing that.
-    denom = abs(early) + abs(late)
-    if denom > 0
-        disc = (abs(early) - abs(late)) / denom
-        p.delta[ch] = clamp(p.delta[ch] + 0.05 * disc * shift * cps,
-                            p.anchor[ch] - 2.0, p.anchor[ch] + 2.0)
-    end
-
-    @printf(p.io, "X %d %d %d %d %.4f %.4f %.6g %.6g %.6g %.6g %.6g %.6g %.6g %.6g %.3f %.5f\n",
-            rec.prn, ch, rec.sample_index, n, rec.code_phase, p.delta[ch],
-            real(rec.prompt), imag(rec.prompt), abs(rec.late), abs(rec.early),
-            real(prompt), imag(prompt), abs(late), abs(early),
-            rec.carrier_hz, rec.code_doppler_hz)
-    p.records += 1
-    p
-end
-
-_xcorr_scalar(x::Complex) = x
-_xcorr_scalar(x::AbstractVector) = x[1]      # multi-antenna: antenna 0 only
-
-# Every record the link drained this chunk. `drain_buffer` is the link's own
-# batch `take!` scratch, so reading it here tees the dump stream without a
-# second consumer on the SPSC ring; records the drain did not refresh are
-# filtered out by the per-channel `sample_index` dedup.
-#
-# The capture step gets the assignment table rather than the link so that its
-# specialisation depends only on the record and tracking-state types — the
-# warm-up's stub SDR then compiles the same method the live run calls — and the
-# correlation step, which is the expensive one, is free of both.
-function xcorr_process!(p::CpuReference, link, track_state)
-    t0 = time()
-    for dump in link.drain_buffer
-        xcorr_capture!(p, link.assignments, track_state, dump)
-    end
-    xcorr_drain_pending!(p)
-    p.cpu_seconds += time() - t0
-    p
-end
-
-# ── Bit-stream instrument ─────────────────────────────────────────────────────
-# `receive`'s payload channel emits once per `pvt_update_interval` (100 ms) but
-# the bit buffer is reset every chunk (2 ms), so the run-script bit log sees
-# ~1/50 of the soft bits — useless for checking preamble spacing or parity.
-# This wraps the correlator source instead: `advance_tracking!` is the one call
-# per chunk that runs after the fold and before `decode` consumes the buffer,
-# so what it sees IS what the decoder gets, bit for bit.
-mutable struct BitLogSource{L}
-    const link::L
-    const io::Union{Nothing,IO}
-    # Per-PRN sink for the raw 1 ms filtered prompts, as interleaved Float32
-    # (re, im). The soft bits alone cannot separate "the 20 ms accumulation is
-    # aligned on the wrong edge" from "the prompts have no phase coherence to
-    # accumulate": with the prompts, both bit phase and coherence can be
-    # measured offline, and the bits re-derived at all 20 edge hypotheses.
-    const prompts::Dict{Int,IO}
-    const prompt_dir::String
-    # The FPGA-vs-CPU probe, or `nothing`. It hangs off the same wrapper because
-    # it needs both halves of one chunk: the raw samples on the way in, and the
-    # records the fold drained on the way out.
-    const xcorr::Union{Nothing,CpuReference}
-    chunks::Int
-    # Every hardware record the fold drained this chunk, as it came off the wire
-    # (`R` lines, in a separate file next to the bit log): the raw 1 ms
-    # accumulators, the device sample span and the replica code phase. The `P`
-    # prompts are what Tracking saw *after* the link coalesced a chunk's dumps
-    # into one record, so post-sync they are 2-8 ms long and unusable for
-    # anything that needs the 1 ms grid — bit-edge re-folds, false-lock
-    # signatures (a 500 Hz Costas alias flips the sign of every consecutive
-    # 1 ms prompt and is invisible at 2 ms), or the exact shape of a stream gap.
-    const records::Union{Nothing,IO}
-    # `drain_dumps!` leaves `drain_buffer` untouched when the ring was empty, so
-    # the newest sample index logged per channel is what tells a fresh drain
-    # from last chunk's buffer seen again.
-    const last_logged::Vector{Int64}
-    # The PRN each hardware channel held at the previous chunk, so a change is
-    # logged once (`A` lines) and a stream gap can be put beside the handover
-    # or re-arm that caused it.
-    const assigned::Vector{Int32}
-end
-
-# The instrument wraps the link, so `process` sees a `BitLogSource` where the
-# dump-gap protection dispatches on a `HardwareCorrelatorLink`. Without this
-# forward every instrumented run — which is every run this issue has measured —
-# silently loses the freeze and decays satellites through a host-side gap.
-GNSSReceiver.is_observation_gap(source::BitLogSource, track_state, group_key, prn) =
-    GNSSReceiver.is_observation_gap(source.link, track_state, group_key, prn)
-
-function GNSSReceiver.advance_tracking!(
-    source::BitLogSource,
-    band_measurements,
-    track_state,
-    band_systems,
-)
-    probe = source.xcorr
-    # Ring the chunk BEFORE the fold: `advance_tracking!` is what advances
-    # `samples_consumed`, so this is the one moment its value still names the
-    # first sample of this chunk.
-    isnothing(probe) || xcorr_append!(
-        probe,
-        Tracking.get_samples(first(values(band_measurements))),
-        source.link.samples_consumed,
-    )
-    track_state = GNSSReceiver.advance_tracking!(
-        source.link,
-        band_measurements,
-        track_state,
-        band_systems,
-    )
-    isnothing(probe) || xcorr_process!(probe, source.link, track_state)
-    log_tracking!(source, track_state)
-end
-
-function log_tracking!(source::BitLogSource, track_state)
-    io = source.io
-    isnothing(io) && return track_state
-    source.chunks += 1
-    link = source.link
-    boundary = link.next_epoch_boundary
-    # Channel table changes: handovers, releases and the noise channel's
-    # re-arms. The noise channel shows up as its decoy PRN changing.
-    for ch in eachindex(link.assignments)
-        assignment = link.assignments[ch]
-        prn_now = ch == link.noise_channel ? -link.noise_prn :
-                  isnothing(assignment) ? Int32(0) : Int32(assignment.prn)
-        if prn_now != source.assigned[ch]
-            println(io, "A ", ch, ' ', source.assigned[ch], ' ', prn_now, ' ', boundary,
-                    ' ', link.samples_consumed)
-            source.assigned[ch] = prn_now
-        end
-    end
-    records = source.records
-    new_dumps = 0
-    for dump in link.drain_buffer
-        GNSSReceiver.is_epoch_strobe(dump) && continue
-        ch = Int(dump.channel)
-        checkbounds(Bool, source.last_logged, ch) || continue
-        dump.output.sample_index <= source.last_logged[ch] && continue
-        source.last_logged[ch] = dump.output.sample_index
-        new_dumps += 1
-        isnothing(records) && continue
-        acc = Tracking.get_accumulators(dump.output.correlator)
-        println(records, "R ", ch, ' ', dump.prn, ' ', dump.output.sample_index,
-                ' ', dump.output.integrated_samples,
-                ' ', round(dump.code_phase; digits = 4),
-                ' ', round(Int, real(acc[1])), ' ', round(Int, imag(acc[1])),
-                ' ', round(Int, real(acc[2])), ' ', round(Int, imag(acc[2])),
-                ' ', round(Int, real(acc[3])), ' ', round(Int, imag(acc[3])))
-    end
-    # One line per chunk: wall clock (ns), the epoch boundary the fold reached,
-    # how many new dumps this chunk drained and the device's newest sample seen.
-    # A steady pipeline drains ~2 dumps per 2 ms chunk; a stall shows up as one
-    # chunk draining tens of dumps followed by a run of chunks draining none,
-    # and the wall clock says whether the stall was the raw stream or the fold.
-    println(io, "T ", time_ns(), ' ', boundary, ' ', new_dumps, ' ', link.latest_sample_index,
-            ' ', link.samples_consumed, ' ', Base.cumulative_compile_time_ns()[1],
-            ' ', Base.gc_num().total_time)
-    gk = GNSSReceiver.signal_group_key(gpsl1)
-    for (prn, sat) in pairs(Tracking.get_sat_states(track_state))
-        signal = Tracking.get_signals(sat)[1]
-        soft_bits = signal.bit_buffer.soft_bits
-        if !isempty(soft_bits)
-            print(io, "B ", prn, ' ', boundary)
-            for b in soft_bits
-                print(io, ' ', b)
-            end
-            println(io)
-        end
-        prompts = Tracking.get_filtered_prompts(signal)
-        if !isempty(prompts)
-            sink = get!(source.prompts, prn) do
-                open(joinpath(source.prompt_dir, "prompts_prn$(prn).f32"), "w")
-            end
-            for p in prompts
-                write(sink, Float32(real(p)), Float32(imag(p)))
-            end
-            println(io, "P ", prn, ' ', boundary, ' ', length(prompts),
-                    ' ', signal.bit_buffer.found,
-                    ' ', signal.bit_buffer.prompt_accumulator_integrated_code_blocks)
-            # The prompt amplitude decaying over tens of seconds is the thing
-            # that stops the decoder, so log what would explain it: the early /
-            # late imbalance (the DLL's own view of how far off the peak the
-            # replica sits) and the two Dopplers plus the absolute code phase
-            # the host is steering with.
-            acc = Tracking.get_accumulators(
-                Tracking.get_last_fully_integrated_correlator(signal),
-            )
-            # Appended after the original fields (the analysis scripts read by
-            # position from the front): the C/N₀ the lock detector sees, and the
-            # bit buffer's polarity and pre-sync window length.
-            cn0 = Tracking.estimate_cn0(track_state, gk, prn, 1)
-            println(io, "D ", prn, ' ', boundary,
-                    ' ', abs(acc[1]), ' ', abs(acc[2]), ' ', abs(acc[3]),
-                    ' ', Tracking.get_carrier_doppler(sat),
-                    ' ', Tracking.get_code_doppler(sat),
-                    ' ', Tracking.get_code_phase(sat),
-                    ' ', round(10log10(ustrip(Hz, Unitful.linear(cn0))); digits = 2),
-                    ' ', signal.bit_buffer.polarity,
-                    ' ', signal.bit_buffer.code_block_buffer_length)
-        end
-    end
-    # ~1 s of chunks: the per-channel record-continuity counters. `lost` is in
-    # device samples of records the host never saw — at 4000 samples per code
-    # block that is directly the number of code blocks this satellite's
-    # navigation bit clock has slipped behind the sample axis. The trailing
-    # field is `rearm`, the same unit for signal the device never integrated
-    # because the channel was re-armed: elapsed time, but nobody's fault.
-    if source.chunks % 500 == 0
-        for ch in eachindex(link.assignments)
-            assignment = link.assignments[ch]
-            isnothing(assignment) && continue
-            println(io, "C ", ch, ' ', assignment.prn, ' ', boundary,
-                    ' ', link.lost_record_samples[ch],
-                    ' ', link.overlapping_record_samples[ch],
-                    ' ', link.last_record_end[ch],
-                    ' ', link.rearm_dead_samples[ch])
-        end
-        println(io, "L ", boundary, ' ', link.lost_record_gaps,
-                ' ', link.stale_dumps, ' ', link.dropped_dumps,
-                ' ', link.skipped_epochs, ' ', link.rearm_gaps)
-        flush(io)
-    end
-    track_state
-end
-
-# `receive(::AbstractHardwareCorrelatorSDR, …)` builds the link internally, so
-# to wrap it we build it here and go in through the sample-channel method.
-function hw_receive(
-    sdr,
-    io,
-    prompt_dir = mktempdir();
-    doppler_update_interval = nothing,
-    feedback_delay_epochs::Integer = 2,
-    acquire_async::Bool = true,
-    xcorr = nothing,
-    records = nothing,
-    kwargs...,
-)
-    link = GNSSReceiver.HardwareCorrelatorLink(
-        sdr;
-        sampling_freq = FS,
-        reference_signal = GNSSReceiver.ranging_signal(gpsl1),
-        doppler_update_interval,
-        feedback_delay_epochs,
-    )
-    n_channels = GNSSReceiver.num_hardware_channels(sdr)
-    source = BitLogSource(
-        link, io, Dict{Int,IO}(), prompt_dir, xcorr, 0,
-        records, fill(typemin(Int64), n_channels), zeros(Int32, n_channels),
-    )
-    BIT_LOG_SOURCE[] = source
-    GNSSReceiver.receive(
-        GNSSReceiver.raw_sample_channel(sdr),
-        gpsl1,
-        FS;
-        correlator_source = source,
-        acquire_async,
-        kwargs...,
-    )
-end
-
-const BIT_LOG_SOURCE = Ref{Any}(nothing)
-const BIT_LOG_WARM = Ref{Union{Nothing,IO}}(devnull)
-
-function append_soft_bits!(bits_log, decoded)
-    for e in decoded
-        isempty(e.soft) || append!(get!(bits_log, e.prn, Float32[]), e.soft)
-    end
-    bits_log
-end
-
-fresh_pvt(pvt, previous_time) =
-    !isnothing(pvt.time) && !isequal(pvt.time, previous_time)
-
-position_with_age(position, age) =
-    @sprintf("%s [age %.1f s]", position_summary(position), age)
-
-my_extract(state) = (
-    data = GNSSReceiver.default_data_of_interest(state),
-    decode = [
-        begin
-            sats = Tracking.get_sat_states(state.track_state, gk)
-            bb = haskey(sats, rss.prn) ?
-                 Tracking.get_signals(sats[rss.prn])[1].bit_buffer : nothing
-            (prn = rss.prn,
-             nbits = something(rss.decoder.num_bits_after_valid_syncro_sequence, -1),
-             tow = Int64(something(rss.decoder.data.TOW, -1)),
-             found = isnothing(bb) ? false : bb.found,
-             nsoft = isnothing(bb) ? -1 : length(bb.soft_bits),
-             soft = isnothing(bb) ? Float32[] : copy(bb.soft_bits))
-        end
-        for (gk, dict) in pairs(state.receiver_sat_states) for rss in dict
-    ],
-)
-
-# ── Status reporting ─────────────────────────────────────────────────────────
-# Every line the harness prints *while the receiver runs*, built here rather
-# than inline in the run loop, so that `warm_reporting` below can pay for their
-# compilation before the stream exists.
-#
-# This is not tidiness. `Printf.format` is compiled per format string on first
-# call, and a `@sprintf` sitting in the run loop is therefore first compiled
-# from inside the fold — where the receiver has a few milliseconds before the
-# device's dump ring overruns. Measured on the Orin (issue #107,
-# `--trace-compile-timing`): the first status line cost 743 ms of `Printf`
-# plus 368 + 153 ms for the two comprehensions that build it, and the
-# `slip_watch` timer another 383 ms on its first tick — 1.6 s of stalls that
-# emptied the ring and cost every satellite the records taken during them.
-# Nothing here is on a hot path; it runs at most a few times a second.
-
-sat_summary(sat_data) = join(
-    [@sprintf("%d:%.0f", k[2], 10log10(ustrip(Hz, Unitful.linear(v.cn0))))
-     for (k, v) in pairs(sat_data)],
-    " ",
-)
-
-decode_summary(decode) = join(
-    [@sprintf("%d:%db%s%s%d", e.prn, e.nbits, e.tow >= 0 ? "*" : "",
-              e.found ? "F" : "-", e.nsoft)
-     for e in decode],
-    " ",
-)
-
-position_summary(position) = let lla = LLAfromECEF(wgs84)(position)
-    @sprintf("%.6f°, %.6f°, %.0fm", lla.lat, lla.lon, lla.alt)
-end
-
-# lag = wall time minus receiver runtime: how far the processing is behind real
-# time. The NCO corrections apply `lag` late, so a growing lag silently opens
-# the loop delay — the PLL stops phase-locking long before anything else looks
-# wrong.
-status_line(t, runtime, sats, dec, pos, missed, free_bytes) = @sprintf(
-    "t=%5.1fs rt=%5.1fs cn0[%s] dec[%s] | %s | missed=%d free=%dG",
-    t, ustrip(u"s", runtime), sats, dec, pos, missed, free_bytes ÷ 2^30)
-
-first_fix_line(t, position, num_sats) = let lla = LLAfromECEF(wgs84)(position)
-    @sprintf(
-        "*** FIRST FIX after %.1f s: lat %.6f° lon %.6f° alt %.1f m (%d sats) ***",
-        t, lla.lat, lla.lon, lla.alt, num_sats)
-end
-
-slip_corrected_line(slip, total) = @sprintf(
-    "host axis slipped %+d samples (%.1f chips) — device_origin re-anchored, %+d total",
-    slip, slip * 0.2558, total)
-
-slip_implausible_line(slip) = @sprintf(
-    "host axis slip %+d samples is implausibly large — not correcting", slip)
-
-# `lost` is records that never reached the host; `rearm` is signal time the
-# device did not integrate at all because the channel was being re-armed. They
-# were one counter until the board showed the open-loop noise channel, re-armed
-# once a second, reporting 252 "lost" holes in a run that lost nothing.
-continuity_summary(link) = join(
-    [string(a.prn, "→ch", ch, " lost=", link.lost_record_samples[ch],
-            "smp rearm=", link.rearm_dead_samples[ch],
-            "smp overlap=", link.overlapping_record_samples[ch], "smp")
-     for (ch, a) in enumerate(link.assignments) if !isnothing(a)],
-    "  ",
-)
-
-link_summary(link) = string(
-    "link: lost_gaps=", link.lost_record_gaps, " rearm_gaps=", link.rearm_gaps,
-    " stale=", link.stale_dumps,
-    " dropped=", link.dropped_dumps, " skipped_epochs=", link.skipped_epochs,
-    " implausible=", link.implausible_dumps)
-
-# Run every one of them once, into `devnull`, on values of exactly the types the
-# live loop will hand them. The dummies are built by hand rather than taken from
-# the warm-up's own payloads because those track nothing: an empty `sat_data`
-# never enters the comprehension body, so nothing inside it would compile.
-function warm_reporting()
-    # Exactly the container and element the receiver's own payload carries: the
-    # comprehensions above compile per container type, so a stand-in `Dict`
-    # would leave them to be compiled live after all. `Dictionaries` is
-    # GNSSReceiver's dependency, not the harness's, hence the qualified name.
-    sat_data = GNSSReceiver.Dictionary(
-        [(:GPSL1CA, 1)],
-        [GNSSReceiver.SatelliteDataOfInterest(45.0u"dBHz", 1.0 + 0.0im, true)],
-    )
-    decode = [(prn = 1, nbits = 300, tow = Int64(42), found = true, nsoft = 20,
-               soft = Float32[0.5])]
-    position = ECEF(4.0e6, 0.5e6, 4.6e6)
-    # Invoke standalone methods: an inlined warm-up alone can leave the
-    # dynamically dispatched call in the live reporting loop uncompiled.
-    Base.invokelatest(sat_summary, sat_data)
-    Base.invokelatest(decode_summary, decode)
-    Base.invokelatest(position_summary, position)
-    Base.invokelatest(position_with_age, position, 1.0)
-    Base.invokelatest(append_soft_bits!, Dict{Int,Vector{Float32}}(), decode)
-    # `Sys.free_memory()` is a `UInt64`; `Printf` compiles per argument type, so
-    # the stand-in has to be one too.
-    Base.invokelatest(status_line, 1.0, 1.0u"s", "1:45", "1:0b-F0", "no fix", 0, UInt64(2)^31)
-    Base.invokelatest(first_fix_line, 1.0, position, 4)
-    print(devnull, slip_corrected_line(16, 16))
-    print(devnull, slip_implausible_line(1 << 20))
-    nothing
-end
-
-mutable struct WarmSDR{C} <: AbstractHardwareCorrelatorSDR
-    const raw::SignalChannel{Complex{Int16},1}
-    const dumps::PipeChannel{CorrelatorDump{C}}
-    const ncos::PipeChannel{NCOUpdate}
-end
-warm_epl(l, p, e) = EarlyPromptLateCorrelator(SVector{3,ComplexF64}(l, p, e), 1)
-const WARM_EPL = typeof(warm_epl(0, 0, 0))
-GNSSReceiver.raw_sample_channel(s::WarmSDR) = s.raw
-GNSSReceiver.correlator_dump_channel(s::WarmSDR) = s.dumps
-GNSSReceiver.nco_update_channel(s::WarmSDR) = s.ncos
-GNSSReceiver.num_hardware_channels(s::WarmSDR) = 5
-GNSSReceiver.assign_channel!(s::WarmSDR, args...; kwargs...) = nothing
-GNSSReceiver.release_channel!(s::WarmSDR, ch) = nothing
-
-function warm_tracking_log!(sdr)
-    # Noise-only acquisition cannot exercise per-satellite logging or bit
-    # export. Use a populated state and real IOStreams, just like the run.
-    estimator = ConventionalAssistedPLLAndDLL(; carrier_loop_filter_bandwidth = 12.0Hz)
-    state = GNSSReceiver.ReceiverState(
-        ((gpsl1,),),
-        (L1 = GNSSReceiver.SampleBuffer(Complex{Int16}, CHUNK),);
-        doppler_estimator = estimator,
-    )
-    sat = GNSSReceiver.create_tracked_sat(
-        (gpsl1,), 1, 0.0, 0.0Hz, Tracking.NumAnts(1), estimator,
-    )
-    gk = GNSSReceiver.signal_group_key(gpsl1)
-    track_state = Base.invokelatest(Tracking.merge_sats, state.track_state, gk, [sat])
-    signal = first(Tracking.get_signals(sat))
-    push!(signal.bit_buffer.soft_bits, 0.5f0)
-    push!(Tracking.get_filtered_prompts(signal), 1000.0 + 10.0im)
-    link = GNSSReceiver.HardwareCorrelatorLink(sdr; sampling_freq = FS, reference_signal = gpsl1)
-    link.assignments[2] = GNSSReceiver.HardwareChannelAssignment(gk, 1, 1)
-    link.next_epoch_boundary = 8000
-    push!(link.drain_buffer, CorrelatorDump(2, 1,
-        CorrelatorOutput(warm_epl(400, 1000, 400), 4000, 4000), 0.0))
-    mktempdir() do dir
-        open(joinpath(dir, "warm.log"), "w") do io
-            source = BitLogSource(link, io, Dict{Int,IO}(), dir, nothing, 499,
-                io, fill(typemin(Int64), 5), zeros(Int32, 5))
-            try
-                Base.invokelatest(log_tracking!, source, track_state)
-            finally
-                foreach(close, values(source.prompts))
-            end
-        end
-    end
-    nothing
-end
-
-function warm_device_calls!(sdr)
-    # Compile the dynamically dispatched vendor methods without executing an
-    # assignment or changing a hardware channel during warm-up.
-    kw = (el_sample_spacing = 1, signal = gpsl1)
-    precompile(Core.kwcall, (typeof(kw), typeof(GNSSReceiver.assign_channel!),
-        typeof(sdr), Int, Int, typeof(FS), typeof(FS), Float64, Int))
-    precompile(GNSSReceiver.release_channel!, (typeof(sdr), Int))
-    precompile(GNSSReceiver.dropped_dump_count!, (typeof(sdr),))
-    precompile(GNSSReceiver.assignment_start_sample, (typeof(sdr), Int))
-    nothing
-end
-
-function warm_up()
-    @info "warming up receive() (hardware-path JIT)…"
-    t0 = time()
-    sdr = WarmSDR{WARM_EPL}(
-        SignalChannel{Complex{Int16},1}(CHUNK, 16),
-        PipeChannel{CorrelatorDump{WARM_EPL}}(1 << 16),
-        PipeChannel{NCOUpdate}(1 << 12),
-    )
-    task = Threads.@spawn begin
-        bufs = [Complex{Int16}.(rand(-100:100, CHUNK, 1), rand(-100:100, CHUNK, 1)) for _ = 1:3]
-        nepoch = CHUNK ÷ 4000
-        for i = 1:200
-            batch = CorrelatorDump{WARM_EPL}[]
-            for e = 1:nepoch
-                base = ((i - 1) * nepoch + e) * 4000
-                for ch = 1:2
-                    push!(batch, CorrelatorDump(ch, ch,
-                        CorrelatorOutput(warm_epl(400.0 + 0im, 1000.0 + 10im, 400.0 + 0im),
-                                         4000, base - 50 + ch),
-                        mod(5.0 + 0.001base, 1023.0)))
-                end
-                push!(batch, epoch_strobe(warm_epl(0, 0, 0), base))
-            end
-            put!(sdr.dumps, batch)
-            put!(sdr.raw, bufs[mod1(i, 3)])
-            while Base.n_avail(sdr.ncos) > 0
-                take!(sdr.ncos)
-            end
-        end
-        close(sdr.raw)
-        close(sdr.dumps)
-    end
-    data = hw_receive(
-        sdr,
-        BIT_LOG_WARM[];
-        records = BIT_LOG_WARM[],
-        # The probe's own kernels are compiled by its constructor; this is for
-        # the wrapper around them, which only exists when the probe does.
-        xcorr = isempty(XCORR_LOG_NAME) ? nothing :
-                CpuReference(devnull, 0, FS_HZ, 5; ring_seconds = 0.05),
-        prns = [1, 2, 3, 4, 5, 7],
-        acq_min_doppler_coverage = ACQ_COVERAGE,
-        acq_coherent_integration_time = 10ms,
-        acq_noncoherent_rounds = 5,
-        max_meas = 2^11,
-        acquire_every = ACQ_EVERY,
-        # Same scheduler as the live call, so its merge path is compiled here
-        # and not during the decode window.
-        acquire_async = ACQ_ASYNC,
-        feedback_delay_epochs = 1,
-        doppler_estimator = ConventionalAssistedPLLAndDLL(;
-            carrier_loop_filter_bandwidth = 12.0Hz,
-        ),
-        code_lock_cn0_threshold = 24.0u"dBHz",
-        processing_threadpool = PROC_POOL,
-        extract = my_extract,
-    )
-    for payload in data
-        # The live reporting loop receives a dynamically typed payload. Compile
-        # its standalone property reads too; inlined warm-up does not cover them.
-        snapshot = Base.invokelatest(getproperty, payload, :data)
-        Base.invokelatest(getproperty, snapshot, :sat_data)
-    end
-    wait(task)
-    warm_tracking_log!(sdr)
-    # The harness's own reporting is part of the pipeline as far as the fold is
-    # concerned: it runs on the same task, between chunks.
-    warm_reporting()
-    # Exercise a successful position solve and the HOW recovery helper after
-    # the complete dependency tree has loaded, before feedback is live.
-    pvt_states = GNSSReceiver._precompile_pvt_states()
-    if !isempty(pvt_states)
-        pvt = Base.invokelatest(GNSSReceiver.calc_pvt, pvt_states; approximate_year = 2026)
-        Base.invokelatest(length, pvt.sats)
-        @assert Base.invokelatest(fresh_pvt, pvt, nothing)
-        @assert !Base.invokelatest(fresh_pvt, pvt, pvt.time)
-    end
-    Base.invokelatest(GNSSReceiver.GNSSDecoder.is_plausible_TOW,
-        UInt64(1), nothing, nothing, Int64(0))
-    # Dynamic signed/unsigned navigation-word arithmetic was the remaining
-    # small decoder specialization in the live trace.
-    Base.invokelatest(+, Int64(0), UInt64(0))
-    @info @sprintf("warm-up done in %.1f s", time() - t0)
-end
-
-# ── 2. Raw stream plumbing ────────────────────────────────────────────────────
-# Tap ring: circular, no per-chunk memmove. `pos` is the producer count at the
-# ring's end; the newest sample sits just behind index mod(pos-1, len)+1.
-mutable struct TapRing
-    const data::Vector{ComplexF32}
-    pos::Int
-end
-TapRing(len) = TapRing(zeros(ComplexF32, len), 0)
-
-function tap_append!(ring::TapRing, chunk::AbstractMatrix{Complex{Int16}})
-    n = length(ring.data)
-    base = mod(ring.pos, n)
-    @inbounds for k = 1:size(chunk, 1)
-        ring.data[mod(base + k - 1, n)+1] = ComplexF32(chunk[k, 1])
-    end
-    ring.pos += size(chunk, 1)
-    ring
-end
-
-# Freshest `m` samples in time order plus the producer count at their end.
-function tap_snapshot(ring::TapRing, lock_, m)
-    lock(lock_) do
-        n = length(ring.data)
-        m <= ring.pos || error("tap ring not yet filled ($(ring.pos) < $m)")
-        m <= n || error("tap ring too small ($n < $m)")
-        out = Vector{ComplexF32}(undef, m)
-        start = ring.pos - m           # 0-based producer index of out[1]
-        @inbounds for k = 1:m
-            out[k] = ring.data[mod(start + k - 1, n)+1]
-        end
-        (out, ring.pos)
-    end
-end
-
-mutable struct RawStream
-    channel::SignalChannel{Complex{Int16},1}
-    tap::TapRing
-    tap_lock::ReentrantLock
-    recorder::Base.Process
-    reader::Task
-end
-
-function start_raw_stream(; tap_seconds = 0.6, capacity_chunks = 4000)
-    run(ignorestatus(`pkill -x m2sdr_record`))
-    sleep(0.3)
-
-    channel = SignalChannel{Complex{Int16},1}(CHUNK, capacity_chunks)
-    tap = TapRing(round(Int, tap_seconds * FS_HZ))
-    tap_lock = ReentrantLock()
-
-    # Read the recorder's stdout through a libuv pipe: reads block the *task*,
-    # not the thread. A FIFO + blocking IOStream instead parks the reader's OS
-    # thread in a raw read ccall, and any Polyester sticky worker pinned to
-    # that thread (acquisition!) starves forever.
-    #
-    # Warm the exact read path on a throwaway process first: the recorder
-    # starts streaming the moment it launches, and every millisecond the
-    # reader spends compiling is DMA ring the startup may overrun. (The sweep
-    # calibration absorbs a lossy start into the measured origin either way.)
-    let warm = open(`head -c 65536 /dev/zero`, "r")
-        read!(warm, Vector{UInt8}(undef, 65536))
-        close(warm)
-    end
-    # Real-time priority for the recorder, if the host will grant it.
-    #
-    # Acquisition >= 2.8 respects Julia's thread pools, removing the former
-    # Polyester opening-scan starvation. Real-time scheduling remains optional
-    # protection against unrelated host load; normal priority is a supported
-    # configuration and is what the on-sky tests use on orin2.
-    rt_prefix = try
-        success(pipeline(`chrt -f 50 true`; stdout = devnull, stderr = devnull)) ?
-        `chrt -f 50` : ``
-    catch
-        ``
-    end
-    isempty(rt_prefix.exec) && @info "recorder runs at normal scheduling priority"
-    recorder = open(`$rt_prefix m2sdr_record -q - 0`, "r")
-
-    # 64 KiB of pipe = ~2 ms of slack at 32 MB/s; any longer reader stall backs
-    # into the kernel ring (65 ms) and the driver drops whole buffers, slipping
-    # the host axis ~262 chips (mod 1023) per 1024-sample buffer — every
-    # handover placed after that lands off-peak. ~1 s of slack removes it.
-    let want = 32 * 1024 * 1024
-        got = -1
-        while want >= 1024 * 1024
-            got = ccall(:fcntl, Cint, (Cint, Cint, Cint),
-                        Base._fd(recorder.out), 1031 #= F_SETPIPE_SZ =#, want)
-            got > 0 && break
-            want >>= 1
-        end
-        got > 0 ? (@info "recorder pipe capacity: $(got >> 20) MiB") :
-                  (@warn "could not grow the recorder pipe (losses will slip the host axis)")
-    end
-
-    reader = Threads.@spawn :interactive begin
-        io = recorder
-        nbuf = capacity_chunks + 2
-        pool = [Matrix{Complex{Int16}}(undef, CHUNK, 1) for _ = 1:nbuf]
-        raw = Vector{UInt8}(undef, CHUNK * 8)
-        idx = 1
-        try
-            while isopen(channel)
-                read!(io, raw)
-                words = reinterpret(Int16, raw)     # 4 × Int16 per sample (2R2T)
-                buf = pool[idx]
-                @inbounds for k = 1:CHUNK
-                    buf[k, 1] = Complex(words[4k-3], words[4k-2])   # antenna 0
-                end
-                lock(tap_lock) do
-                    tap_append!(tap, buf)
-                end
-                put!(channel, buf)
-                Threads.atomic_add!(RAW_CHUNKS, 1)
-                idx = mod1(idx + 1, nbuf)
-            end
-        catch e
-            e isa EOFError || e isa InvalidStateException || rethrow()
-        finally
-            close(channel)
-            close(io)
-        end
-    end
-    Base.errormonitor(reader)
-    RawStream(channel, tap, tap_lock, recorder, reader)
-end
-
-acq_snapshot(stream::RawStream, noncoherent) =
-    tap_snapshot(stream.tap, stream.tap_lock, 10 * 4000 * noncoherent)
-
-function acquire_prns(stream::RawStream, prns; noncoherent = 4)
-    signal, host_end = acq_snapshot(stream, noncoherent)
-    results = acquire(gpsl1, signal, FS, prns;
-                      min_doppler_coverage = ACQ_COVERAGE,
-                      num_coherently_integrated_code_periods = 10,
-                      num_noncoherent_accumulations = noncoherent)
-    # Acquired code phases are referenced to the start of the signal buffer.
-    (results, host_end - length(signal))
-end
-
-
-
-
-
-
-
-# ── Main ──────────────────────────────────────────────────────────────────────
-function main()
-    # Include compilation and GC counters in the T records so latency can be
-    # measured without locally patching the harness before each hardware run.
-    Base.cumulative_compile_timing(true)
-    # Redirected stderr is fully buffered; flush it so the log is observable
-    # live. A Timer callback that throws kills the timer silently — never let
-    # that happen to the only thing making the log visible.
-    flusher = Timer(1; interval = 2) do _
-        try
-            flush(stderr)
-            flush(stdout)
-        catch
-        end
-    end
-
-    warm_up()
-
-    # Latch the origin BEFORE streaming: the bank counter only advances on
-    # samples DMA0 accepts, so its pre-stream value is exactly the device index
-    # of the stream's first sample.
-    csr_probe = GNSSM2SDR.LiteXCSR(CSR_CSV)
-    bank_probe = GNSSM2SDR.GNSSBank(csr_probe; fs = FS)
-    origin_prelatch = sample_count(bank_probe)
-    close(csr_probe)
-    @info "pre-stream device counter: $origin_prelatch"
-
-    stream = start_raw_stream()
-    @info "raw stream started"
-    let last = Ref(0)
-        global stream_watchdog = Timer(5; interval = 5) do _
-            c = RAW_CHUNKS[]
-            rate = (c - last[]) / 5
-            alive = process_running(stream.recorder)
-            (rate < 250 || !alive) &&
-                @warn "raw stream unhealthy" chunks_per_s = rate recorder_alive = alive
-            last[] = c
-        end
-    end
-
-    # Only drive 8 of the 20 hardware channels: every channel dumps at ~1 kHz
-    # whether assigned or not (there is no per-channel enable), so the CSR
-    # poller's ioctl load scales with the bank size — and 8 is plenty for a fix.
-    sdr = M2SDRCorrelator(CSR_CSV, stream.channel; fs = FS, n_channels = 8)
-    warm_device_calls!(sdr)
-    @info "gateware exposes $(GNSSReceiver.num_hardware_channels(sdr)) channels"
-
-    # Poll dumps over CSR (the doc'd bring-up path: DMA1's IRQ cadence is far
-    # too slow for a 1 kHz loop until the driver gains per-channel buffers).
-    # DMA1 dump source: with the rebuilt driver (DMA_BUFFER_PER_IRQ = 1) a
-    # buffer completes and interrupts every 64 records — at a 16 kHz strobe
-    # that is ~2-3 ms of dump latency, inside the 8 Hz PLL's delay budget, and
-    # the ring makes the dump stream LOSSLESS. CSR polling could never be:
-    # its depth-1 dump registers turned every scheduler hiccup into missed
-    # dumps, and one missed dump per subframe breaks that subframe's parity.
-    start!(sdr; dump_source = :dma, epoch_period = 50)
-    sdr.device_origin = origin_prelatch
-
-    # Host-axis slip watch, and the correction it feeds back.
-    #
-    # A handover describes the satellite at `valid_at_sample`, a count of raw
-    # samples the *host* has consumed, and GNSSM2SDR turns that into a device
-    # sample by adding a constant `device_origin` latched once at start-up. The
-    # constant is only right while the two counters advance together — and they
-    # do not: DMA0 drops a buffer now and then, so the device counts samples the
-    # host never receives and the offset creeps. Nothing notices, because
-    # everything already tracking is bookkept on the device's own axis
-    # (`CorrelatorDump.sample_index`, and the link's code-phase anchor). Only
-    # *new* handovers are affected, and they are affected fatally: the code
-    # phase is off by the accumulated loss, 0.2558 chips per sample, so a
-    # satellite acquired after ~4 samples of creep is handed a replica off its
-    # own correlation peak and never locks. Measured live: +5431 samples over
-    # 270 s = 366 chips, after which no newly acquired satellite ever locked and
-    # the receiver sat on the three it already had — one short of a fix.
-    #
-    # Measuring the creep needs care. `infl = (device counter − origin) −
-    # delivered host samples` rises by one for every sample the device counts and
-    # drops by `CHUNK` each time the producer hands a chunk over, so it sweeps
-    # `[creep, creep + CHUNK)` and equals the creep only just after a hand-over.
-    # Read at an arbitrary instant it is the creep plus up to a whole chunk of
-    # phase, and a minimum over a handful of sparse reads is still worth about
-    # `CHUNK/n` — which is why the old watch could only justify a 1200-sample
-    # (300-chip) alarm threshold and could not have driven a correction. Waiting
-    # for the hand-over removes the term outright and leaves the creep itself,
-    # good to the few samples the two reads are apart.
-    creep_now = function (timeout_s = 0.05)
-        seen = RAW_CHUNKS[]
-        deadline = time() + timeout_s
-        while RAW_CHUNKS[] == seen
-            time() > deadline && return nothing
-            yield()
-        end
-        chunks = RAW_CHUNKS[]
-        (sample_count(sdr.bank) - sdr.device_origin) - Int64(chunks) * CHUNK
-    end
-
-    # Armed only once calibration has finished rewriting `device_origin` — see
-    # `arm_slip_watch!` below. The `+= consumed_before_receive` above moves the
-    # origin deliberately, and a baseline taken across that move would read the
-    # deliberate part as creep and correct it away.
-    slip_baseline = Ref{Union{Nothing,Int64}}(nothing)
-    let corrected = Ref{Int64}(0)
-        # The tick, as a named function rather than the timer's own closure, so
-        # `arm_slip_watch!` can run it once while nothing depends on the fold.
-        # As a timer body it was first compiled on the ten-second tick — 383 ms
-        # of JIT on the receiver's task, with the device's dump ring unattended
-        # throughout (issue #107).
-        slip_tick = function ()
-            try
-                isnothing(slip_baseline[]) && return
-                # A hand-over landing between the two reads shows up as one chunk
-                # too low, i.e. a negative slip, which is discarded below; taking
-                # the smallest of a few reads keeps the estimate on the floor.
-                samples = filter(!isnothing, [creep_now() for _ = 1:5])
-                isempty(samples) && return
-                creep = minimum(samples)
-                slip = creep - slip_baseline[]
-                # One sample is 0.256 chips and a handover wants the code phase
-                # inside about half a chip, so there is no point waiting for the
-                # damage to be visible — but stay clear of the read race, which
-                # can only ever be a whole chunk.
-                # Correcting the origin by `slip` lowers the next `creep` reading
-                # by exactly `slip`, i.e. back onto the baseline, so the baseline
-                # itself stays put.
-                if 8 < slip < 4 * CHUNK
-                    sdr.device_origin += slip
-                    corrected[] += slip
-                    @info slip_corrected_line(slip, corrected[])
-                elseif slip >= 4 * CHUNK
-                    @warn slip_implausible_line(slip)
-                end
-            catch
-            end
-        end
-
-        global arm_slip_watch! = function ()
-            samples = filter(!isnothing, [creep_now() for _ = 1:5])
-            slip_baseline[] = isempty(samples) ? nothing : minimum(samples)
-            # Compiles the tick (and `creep_now`) here. The baseline was taken
-            # from the same reads, so this tick measures a slip of about zero
-            # and corrects nothing.
-            slip_tick()
-        end
-
-        # Compile the Timer wrapper immediately, while the baseline is unset
-        # and before acquisition hands any satellites to the tracking loops.
-        global slip_watch = Timer(0; interval = 10) do _
-            slip_tick()
-        end
-    end
-
-    # Keep the channel drained during calibration and count what we eat.
-    drain_running = Ref(true)
-    drained = Threads.Atomic{Int}(0)
-    # Batch-drain: PipeChannel's single-item blocking take! parks in a
-    # sleep-poll (~10 ms granularity), capping a one-at-a-time drainer at
-    # ~100 chunks/s — an order of magnitude below the stream rate. The backlog
-    # that builds then makes `valid_at` minutes stale by receiver start, and
-    # the handover propagation degrades with it.
-    drainer = Threads.@spawn :interactive begin
-        while drain_running[]
-            n = Base.n_avail(stream.channel.channel)
-            if n == 0
-                sleep(0.002)
-                continue
-            end
-            take!(stream.channel, n)
-            Threads.atomic_add!(drained, n * CHUNK)
-        end
-    end
-    Base.errormonitor(drainer)
-
-    # Wait for the tap ring to actually fill (a fixed nap raced a slow reader
-    # start and threw "tap ring not yet filled").
-    let deadline = time() + 30
-        while stream.tap.pos < length(stream.tap.data) && time() < deadline
-            sleep(0.1)
-        end
-        stream.tap.pos >= length(stream.tap.data) ||
-            error("raw stream did not fill the tap ring within 30 s — is m2sdr_record streaming?")
-    end
-    results, _ = acquire_prns(stream, collect(1:32); noncoherent = 10)
-    visible = sort(filter(r -> r.CN0 > RX_FLOOR_DBHZ, results); by = r -> -r.CN0)
-    @info "acquired $(length(visible)) satellites above $(RX_FLOOR_DBHZ) dBHz:"
-    for r in visible
-        @info @sprintf("  PRN %2d  CN0 %.1f dBHz  doppler %+7.0f Hz",
-                       r.prn, r.CN0, ustrip(Hz, r.carrier_doppler))
-    end
-
-    # No origin calibration. `device_origin` is the bank's sample counter latched
-    # immediately before the stream started, and that latch is exact up to the
-    # samples DMA0 drops between the latch and the first delivered buffer.
-    #
-    # This used to be followed by a code-phase sweep that probed ±6 chips of
-    # handover offset and corrected the origin by the peak. Across 36 runs on
-    # this board every *accepted* peak (≥3× the scan floor) sat at 0.00 or
-    # ±0.25 chips — the sweep's own step size — so it only ever confirmed what
-    # the latch already had; the handful of multi-chip results all came from the
-    # one run whose peaks never cleared the acceptance gate, i.e. they were noise
-    # picks. Acquisition's code phase is good to well under a chip on its own,
-    # the residual is inside the DLL's pull-in, and the sweep cost ~2.9 s of the
-    # ~8 s start-up. The `slip_watch` timer below still re-anchors the origin
-    # during the run, which is where the mapping genuinely does drift.
-
-    # Hand the stream to the receiver: stop draining, account for what we ate,
-    # flush stale records.
-    drain_running[] = false
-    # Unblock the drainer if it is parked in take! by feeding it nothing — it
-    # exits after its current take!, which the reader keeps satisfying.
-    wait_t0 = time()
-    while !istaskdone(drainer) && time() - wait_t0 < 3
-        sleep(0.01)
-    end
-    consumed_before_receive = drained[]
-    sdr.device_origin += consumed_before_receive
-    while Base.n_avail(sdr.dumps) > 0
-        take!(sdr.dumps)
-    end
-    # This is the last deliberate move of the origin, so the host-axis creep can
-    # now be baselined: from here any growth is DMA0 loss and nothing else.
-    arm_slip_watch!()
-    @info "starting receiver ($(consumed_before_receive) samples consumed pre-receive)"
-
-    # Decode progress alongside the default payload: bits since preamble sync
-    # and the decoded TOW per satellite — the visibility that separates "loop
-    # works, bits scrambled" from "waiting on subframes".
-    # Both decoder fields are Union{Nothing,Int} — an unguarded conversion
-    # throws inside the processing task and silently closes the data channel.
-    # `found`/`nsoft` expose Tracking's bit-buffer state: whether bit sync ever
-    # happened and how many soft bits this chunk carried to the decoder.
-
-    # Search the whole constellation on periodic scans. Restricting the plan
-    # to the opening scan's detections permanently excluded satellites hidden
-    # by a temporary fade (and capped a two-satellite start at two forever).
-    # Try the initial detections first, strongest first, then the remaining
-    # PRNs; CFAR still gates every handover and the link limits channel use.
-    receiver_prns = [Int(r.prn) for r in visible]
-    append!(receiver_prns, setdiff(1:32, receiver_prns))
-    @info "receiver PRN search order: $receiver_prns"
-    bit_log = isempty(BIT_LOG_NAME) ? nothing :
-              open(joinpath(dirname(@__FILE__), BIT_LOG_NAME), "w")
-    isnothing(bit_log) ||
-        @info "logging every soft bit and 1 ms prompt to $(BIT_LOG_NAME)"
-    # Every raw hardware record beside it (`R` lines, ~120 B per record: five
-    # satellites at 1 kHz write ~35 MB per minute).
-    records_log = isnothing(bit_log) ? nothing :
-                  open(joinpath(dirname(@__FILE__), BIT_LOG_NAME * ".records"), "w")
-    isnothing(records_log) ||
-        @info "logging every hardware record to $(BIT_LOG_NAME).records"
-    # The probe's device axis is `sdr.device_origin` as it stands now — the
-    # calibrated origin plus everything the pre-receiver drainer ate. It is
-    # fixed for the rest of the run, which is exactly what makes a raw sample
-    # addressable by the device index a record reports.
-    xcorr_log = isempty(XCORR_LOG_NAME) ? nothing :
-                open(joinpath(dirname(@__FILE__), XCORR_LOG_NAME), "w")
-    probe = isnothing(xcorr_log) ? nothing :
-            CpuReference(xcorr_log, sdr.device_origin, FS_HZ,
-                         GNSSReceiver.num_hardware_channels(sdr);
-                         prns = XCORR_PRNS, every = XCORR_EVERY)
-    isnothing(probe) || @info string(
-        "FPGA-vs-CPU probe → ", XCORR_LOG_NAME, " (origin ", sdr.device_origin,
-        ", PRNs ", isempty(XCORR_PRNS) ? "all" : join(sort!(collect(XCORR_PRNS)), ","),
-        ", every ", XCORR_EVERY, " records)")
-    # GC pauses stop the CSR poller with the rest of the world; each pause is a
-    # burst of missed dumps = bit slips that scramble the decoders. With 55 GB
-    # free and this run capped at minutes, trading memory for a pause-free
-    # decode window is the right deal — with a valve, re-enabled if RAM runs low.
-    # The full collection runs BEFORE the receiver's processing task exists: on
-    # the heap the calibration scans leave behind it takes seconds, and measured
-    # after `hw_receive` it was the receiver's first stall (3.6 s of held NCO
-    # words on every channel, `T` lines in issue #107).
-    GC.gc(true)
-    GC.enable(false)
-    data = hw_receive(
-        sdr,
-        bit_log,
-        dirname(@__FILE__);
-        xcorr = probe,
-        records = records_log,
-        prns = receiver_prns,
-        acq_min_doppler_coverage = ACQ_COVERAGE,
-        acq_coherent_integration_time = 10ms,
-        # Five rounds rather than three: the satellites that decide whether
-        # there is a fourth at all sit at 25-32 dBHz, where detection
-        # sensitivity is what stands between "searched" and "handed over",
-        # and the scan runs off the processing task anyway.
-        acq_noncoherent_rounds = 5,
-        max_meas = 2^11,
-        # The first acquisition fires as soon as the buffer fills (the timer
-        # starts at -Inf). With asynchronous acquisition the periodic rescan
-        # costs the pipeline nothing, so it runs at a cadence that can actually
-        # recover a lost satellite instead of being switched off.
-        acquire_every = ACQ_EVERY,
-        acquire_async = ACQ_ASYNC,
-        # Loop-delay management: the correction from epoch k reaches the NCO
-        # ~τ = (fold latency + feedback delay) later. The default 18 Hz PLL at
-        # τ ≈ 4-5 ms sits in the delay-instability zone (BL·τ ≈ 0.08): power
-        # and frequency hold, phase never locks, no bits decode. Cut both
-        # knobs: 1-epoch apply (a late commit just applies immediately) and an
-        # retuned PLL below. Runtime compilation must also be warmed: a
-        # buffered dump stream does not protect the loop's feedback deadline.
-        feedback_delay_epochs = 1,
-        doppler_estimator = ConventionalAssistedPLLAndDLL(;
-            carrier_loop_filter_bandwidth = 12.0Hz,
-        ),
-        # Hold a satellite through a fade rather than dropping it at the
-        # default 30 dBHz. Losing lock is not free here: reacquisition resets
-        # symbol timing and TOW. Cached ephemeris fields can shorten recovery,
-        # but a fix still needs four satellites with valid timing at once.
-        # Measured on sky: the
-        # 31-40 dBHz satellites breathe several dB either side of the threshold
-        # and cycle in and out every few tens of seconds, which is exactly long
-        # enough to never finish. The floor still has to sit above where the
-        # loops themselves let go — 24 dBHz is ~6 dB of fade margin on a 1 ms
-        # discriminator, not an invitation to track noise.
-        code_lock_cn0_threshold = 24.0u"dBHz",
-        processing_threadpool = PROC_POOL,
-        extract = my_extract,
-    )
-
-    gc_off = true
-
-
-    t0 = time()
-    last_print = 0.0
-    first_fix = nothing
-    fix_time = 0.0
-    n_solutions = 0
-    previous_pvt_time = nothing
-    last_fresh_fix = 0.0
-    bits_log = Dict{Int,Vector{Float32}}()
-    for payload in data
-        d = payload.data
-        append_soft_bits!(bits_log, payload.decode)
-        t = time() - t0
-        has_fix = d.pvt.time !== nothing
-        if fresh_pvt(d.pvt, previous_pvt_time)
-            previous_pvt_time = d.pvt.time
-            last_fresh_fix = t
-            n_solutions += 1
-            if isnothing(first_fix)
-                first_fix = d.pvt
-                fix_time = t
-                @info first_fix_line(t, d.pvt.position, length(d.pvt.sats))
-            end
-        end
-        if t - last_print >= 5.0
-            last_print = t
-            @info status_line(
-                t,
-                d.runtime,
-                sat_summary(d.sat_data),
-                decode_summary(payload.decode),
-                has_fix ? position_with_age(d.pvt.position, t - last_fresh_fix) : "no fix",
-                sdr.missed_csr_dumps,
-                Sys.free_memory(),
-            )
-            # Memory valve: a paused decode beats an OOM kill.
-            if gc_off && Sys.free_memory() < 8 * 2^30
-                @warn "low memory — re-enabling GC"
-                GC.enable(true)
-                gc_off = false
-            end
-        end
-        (!isnothing(first_fix) && t - fix_time > RUN_AFTER_FIX) && break
-        t > MAX_SECONDS && break
-    end
-
-    @info "stopping"
-    # Both watchdogs measure a *running* receiver. Left armed through the
-    # shutdown below they report on a stream the harness has itself just torn
-    # down — the raw watchdog in particular then prints `recorder_alive = false`
-    # about a recorder this function killed two lines earlier, which reads
-    # exactly like a mid-run crash and cost real time to chase.
-    close(stream_watchdog)
-    close(slip_watch)
-    # Closing the output asks receive() to stop; its finally block joins the
-    # acquisition workers and closes the sample channel. Keep log sinks and
-    # the FPGA services alive until that processing task has stopped writing.
-    close(data)
-    wait(stream.reader)
-    process_running(stream.recorder) && kill(stream.recorder)
-    wait(stream.recorder)
-    stop!(sdr)
-    isnothing(sdr.reader) || wait(sdr.reader)
-    isnothing(sdr.writer) || wait(sdr.writer)
-    let link = BIT_LOG_SOURCE[].link
-        @info "record continuity: " * continuity_summary(link)
-        @info link_summary(link)
-    end
-    if !isnothing(probe)
-        @info string(
-            "FPGA-vs-CPU probe: ", probe.records, " records, mixing sign ", probe.sign,
-            ", CPU-side delta ", join([@sprintf("%.3f", d) for d in probe.delta], "/"),
-            " chips; ", probe.skipped_old, " records rolled out of the ring, ",
-            probe.dropped_pending, " dropped from the wait queue, ",
-            probe.unaligned, " unaligned; deepest record→sample lag ",
-            probe.max_wait, " samples; ",
-            @sprintf("%.1f", probe.cpu_seconds), " s of CPU")
-        close(xcorr_log)
-    end
-    isnothing(bit_log) || close(bit_log)
-    isnothing(records_log) || close(records_log)
-    foreach(close, values(BIT_LOG_SOURCE[].prompts))
-    for (prn, v) in bits_log
-        open(joinpath(dirname(@__FILE__), "bits_prn$(prn).f32"), "w") do io
-            write(io, v)
-        end
-    end
-    @info "soft-bit totals: " * join(["$k:$(length(v))" for (k, v) in bits_log], " ")
-    gc_off && GC.enable(true)
-
-    if isnothing(first_fix)
-        @error "no position fix obtained"
-        # Let any errormonitor'd task failure reach the log before exiting.
-        sleep(3)
-        flush(stderr)
-        exit(1)
-    end
-    lla = LLAfromECEF(wgs84)(first_fix.position)
-    @info @sprintf("PVT returned: first solution at t=%.1fs → %.6f° %.6f° %.1f m; %d fresh timestamp updates (accuracy not validated)",
-                   fix_time, lla.lat, lla.lon, lla.alt, n_solutions)
-end
-
-main()
